+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/create/reproof/CreateReProofUseCase.java
@@ -1,0 +1,8 @@
+package me.jinjjahalgae.domain.proof.usecase.create.reproof;
+
+import me.jinjjahalgae.domain.proof.usecase.create.proof.ProofCreateRequest;
+
+public interface CreateReProofUseCase {
+
+    void execute(ProofCreateRequest request, Long proofId, Long userId);
+}
