--- src/main/java/me/jinjjahalgae/domain/proof/usecase/createreproof/CreateReProofUseCase.java
@@ -1,8 +1,0 @@
-package me.jinjjahalgae.domain.proof.usecase.createreproof;
-
-import me.jinjjahalgae.domain.proof.usecase.createproof.ProofCreateRequest;
-
-public interface CreateReProofUseCase {
-
-    void execute(ProofCreateRequest request, Long proofId, Long userId);
-}
