+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/get/detail/GetProofDetailUseCaseImpl.java
@@ -1,8 +1,9 @@
-package me.jinjjahalgae.domain.proof.usecase.getawaitproof;
+package me.jinjjahalgae.domain.proof.usecase.get.detail;
 
 import lombok.RequiredArgsConstructor;
 import me.jinjjahalgae.domain.participation.repository.ParticipationRepository;
 import me.jinjjahalgae.domain.proof.entities.Proof;
+import me.jinjjahalgae.domain.proof.mapper.ProofImageMapper;
 import me.jinjjahalgae.domain.proof.mapper.ProofMapper;
 import me.jinjjahalgae.domain.proof.repository.ProofRepository;
 import me.jinjjahalgae.global.exception.ErrorCode;
@@ -13,31 +14,29 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class GetAwaitProofUseCaseImpl implements GetAwaitProofUseCase {
-
+public class GetProofDetailUseCaseImpl implements GetProofDetailUseCase {
     private final ProofRepository proofRepository;
     private final ParticipationRepository participationRepository;
 
     @Override
     @Transactional(readOnly = true)
-    public List<ProofAwaitResponse> execute(Long contractId, Long userId) {
+    public ProofDetailResponse execute(Long proofId, Long userId) {
+        // 인증 조회 (이미지 + 피드백)
+        Proof proof = proofRepository.findByIdWithProofImagesAndFeedbacks(proofId)
+                .orElseThrow(() -> ErrorCode.PROOF_NOT_FOUND.domainException(proofId + "에 대한 인증이 존재하지 않습니다."));
+
         // 유저가 계약의 참여자인지 확인
-        boolean isUserParticipate = participationRepository.existsByContractIdAndUserId(contractId, userId);
+        boolean isUserParticipate = participationRepository.existsByContractIdAndUserId(proof.getContractId(), userId);
 
         // 유저가 계약의 참여자가 아닐 경우 예외
         if (!isUserParticipate) {
             throw ErrorCode.ACCESS_DENIED.domainException("계약에 대한 접근 권한이 없습니다.");
         }
 
-        // 인증이 대기 상태이고 감독자가 처리하지 않은 계약의 인증 id들을 조회
-        List<Long> proofIds = proofRepository.findPendingProofIdsWithoutUserFeedback(contractId, userId);
+        // 이미지 순서대로 key만 추출
+        List<String> images = ProofImageMapper.toListResponse(proof.getProofImages());
 
-        // 인증 id로 인증 이미지와 함께 조회
-        List<Proof> proofs = proofRepository.findProofsWithProofImagesByIds(proofIds);
-
-        // dto 응답 반환
-        return proofs.stream()
-                .map(ProofMapper::toAwaitResponse)
-                .toList();
+        // 이미지와 인증 정보와 피드백을 합친 응답 생성
+        return ProofMapper.toDetailResponse(proof, images);
     }
 }
