+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/get/detail/ProofDetailResponse.java
@@ -1,4 +1,4 @@
-package me.jinjjahalgae.domain.proof.usecase.getproofdetail;
+package me.jinjjahalgae.domain.proof.usecase.get.detail;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import me.jinjjahalgae.domain.feedback.dto.FeedbackResponse;
