+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/get/recent/GetRecentProofUseCase.java
@@ -1,4 +1,4 @@
-package me.jinjjahalgae.domain.proof.usecase.getrecentproof;
+package me.jinjjahalgae.domain.proof.usecase.get.recent;
 
 import java.util.List;
