+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/getlist/contractorlist/ContractorProofListResponse.java
@@ -1,4 +1,4 @@
-package me.jinjjahalgae.domain.proof.usecase.getcontractorprooflist;
+package me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist;
 
 import io.swagger.v3.oas.annotations.media.Schema;
