+++ src/main/java/me/jinjjahalgae/presentation/api/ProofController.java
@@ -2,19 +2,19 @@ package me.jinjjahalgae.presentation.api;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import me.jinjjahalgae.domain.proof.usecase.createproof.ProofCreateRequest;
-import me.jinjjahalgae.domain.proof.usecase.createproof.CreateProofUseCase;
-import me.jinjjahalgae.domain.proof.usecase.createreproof.CreateReProofUseCase;
-import me.jinjjahalgae.domain.proof.usecase.getawaitproof.GetAwaitProofUseCase;
-import me.jinjjahalgae.domain.proof.usecase.getawaitproof.ProofAwaitResponse;
-import me.jinjjahalgae.domain.proof.usecase.getcontractorprooflist.ContractorProofListResponse;
-import me.jinjjahalgae.domain.proof.usecase.getcontractorprooflist.GetContractorProofListUseCase;
-import me.jinjjahalgae.domain.proof.usecase.getproofdetail.GetProofDetailUseCase;
-import me.jinjjahalgae.domain.proof.usecase.getproofdetail.ProofDetailResponse;
-import me.jinjjahalgae.domain.proof.usecase.getrecentproof.GetRecentProofUseCase;
-import me.jinjjahalgae.domain.proof.usecase.getrecentproof.ProofRecentResponse;
-import me.jinjjahalgae.domain.proof.usecase.getsupervisorprooflist.GetSupervisorProofListUseCase;
-import me.jinjjahalgae.domain.proof.usecase.getsupervisorprooflist.SupervisorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.create.proof.ProofCreateRequest;
+import me.jinjjahalgae.domain.proof.usecase.create.proof.CreateProofUseCase;
+import me.jinjjahalgae.domain.proof.usecase.create.reproof.CreateReProofUseCase;
+import me.jinjjahalgae.domain.proof.usecase.get.await.GetAwaitProofUseCase;
+import me.jinjjahalgae.domain.proof.usecase.get.await.ProofAwaitResponse;
+import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.ContractorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.GetContractorProofListUseCase;
+import me.jinjjahalgae.domain.proof.usecase.get.detail.GetProofDetailUseCase;
+import me.jinjjahalgae.domain.proof.usecase.get.detail.ProofDetailResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.recent.GetRecentProofUseCase;
+import me.jinjjahalgae.domain.proof.usecase.get.recent.ProofRecentResponse;
+import me.jinjjahalgae.domain.proof.usecase.getlist.supervisorlist.GetSupervisorProofListUseCase;
+import me.jinjjahalgae.domain.proof.usecase.getlist.supervisorlist.SupervisorProofListResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 import me.jinjjahalgae.global.security.jwt.CustomJwtPrincipal;
 import me.jinjjahalgae.presentation.api.docs.proof.ProofControllerDocs;
