+++ src/main/java/me/jinjjahalgae/presentation/api/docs/proof/GetContractorProofListSwaggerResponse.java
@@ -1,6 +1,6 @@
 package me.jinjjahalgae.presentation.api.docs.proof;
 
-import me.jinjjahalgae.domain.proof.usecase.getcontractorprooflist.ContractorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.ContractorProofListResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 
 public class GetContractorProofListSwaggerResponse extends CommonResponse<ContractorProofListResponse> {
