+++ src/main/java/me/jinjjahalgae/presentation/api/docs/proof/GetProofDetailSwaggerResponse.java
@@ -1,6 +1,6 @@
 package me.jinjjahalgae.presentation.api.docs.proof;
 
-import me.jinjjahalgae.domain.proof.usecase.getproofdetail.ProofDetailResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.detail.ProofDetailResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 
 public class GetProofDetailSwaggerResponse extends CommonResponse<ProofDetailResponse> {
