+++ src/main/java/me/jinjjahalgae/presentation/api/docs/proof/ProofControllerDocs.java
@@ -8,12 +8,12 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import me.jinjjahalgae.domain.proof.usecase.createproof.ProofCreateRequest;
-import me.jinjjahalgae.domain.proof.usecase.getawaitproof.ProofAwaitResponse;
-import me.jinjjahalgae.domain.proof.usecase.getcontractorprooflist.ContractorProofListResponse;
-import me.jinjjahalgae.domain.proof.usecase.getproofdetail.ProofDetailResponse;
-import me.jinjjahalgae.domain.proof.usecase.getrecentproof.ProofRecentResponse;
-import me.jinjjahalgae.domain.proof.usecase.getsupervisorprooflist.SupervisorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.create.proof.ProofCreateRequest;
+import me.jinjjahalgae.domain.proof.usecase.get.await.ProofAwaitResponse;
+import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.ContractorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.detail.ProofDetailResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.recent.ProofRecentResponse;
+import me.jinjjahalgae.domain.proof.usecase.getlist.supervisorlist.SupervisorProofListResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 import me.jinjjahalgae.global.security.jwt.CustomJwtPrincipal;
 import org.springframework.web.ErrorResponse;
