+++ src/test/java/me/jinjjahalgae/domain/proof/usecase/createproof/CreateProofUseCaseUnitTest.java
@@ -7,6 +7,8 @@ import me.jinjjahalgae.domain.proof.entities.ProofImage;
 import me.jinjjahalgae.domain.proof.enums.ProofStatus;
 import me.jinjjahalgae.domain.proof.repository.ProofImageRepository;
 import me.jinjjahalgae.domain.proof.repository.ProofRepository;
+import me.jinjjahalgae.domain.proof.usecase.create.proof.CreateProofUseCaseImpl;
+import me.jinjjahalgae.domain.proof.usecase.create.proof.ProofCreateRequest;
 import me.jinjjahalgae.domain.proof.util.ProofTestUtil;
 import me.jinjjahalgae.global.exception.AppException;
 import org.junit.jupiter.api.BeforeEach;
