+++ src/test/java/me/jinjjahalgae/domain/proof/usecase/getawaitproof/GetAwaitProofUseCaseUnitTest.java
@@ -4,6 +4,8 @@ import me.jinjjahalgae.domain.participation.repository.ParticipationRepository;
 import me.jinjjahalgae.domain.proof.entities.Proof;
 import me.jinjjahalgae.domain.proof.enums.ProofStatus;
 import me.jinjjahalgae.domain.proof.repository.ProofRepository;
+import me.jinjjahalgae.domain.proof.usecase.get.await.GetAwaitProofUseCaseImpl;
+import me.jinjjahalgae.domain.proof.usecase.get.await.ProofAwaitResponse;
 import me.jinjjahalgae.domain.proof.util.ProofTestUtil;
 import me.jinjjahalgae.global.exception.AppException;
 import org.junit.jupiter.api.BeforeEach;
