+++ src/test/java/me/jinjjahalgae/domain/proof/usecase/getproofdetail/GetProofDetailUseCaseUnitTest.java
@@ -4,6 +4,8 @@ import me.jinjjahalgae.domain.participation.repository.ParticipationRepository;
 import me.jinjjahalgae.domain.proof.entities.Proof;
 import me.jinjjahalgae.domain.proof.enums.ProofStatus;
 import me.jinjjahalgae.domain.proof.repository.ProofRepository;
+import me.jinjjahalgae.domain.proof.usecase.get.detail.GetProofDetailUseCaseImpl;
+import me.jinjjahalgae.domain.proof.usecase.get.detail.ProofDetailResponse;
 import me.jinjjahalgae.domain.proof.util.ProofTestUtil;
 import me.jinjjahalgae.global.exception.AppException;
 import org.junit.jupiter.api.BeforeEach;
